requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.11.12
click==8.3.1
discord.py==2.6.4
fastapi==0.124.2
frozenlist==1.8.0
h11==0.16.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
romkan @ git+https://github.com/Bel-Shazzar/python-romkan.git@411ad7aadc3d39a46ee271ffbee7697f0a25e1e7
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
yarl==1.22.0
